Widen Schnorr-Euchner lattice coords to i128
Reduced-basis SE coordinates span ~√κ(Q) with κ(Q) ≈ 16/ε⁴, crossing
2^63 near ε=1e-10 (inner shell k≈43) while the reconstructed point x
stays ~2^21. At i64 the per-depth center rounded through to_i64()
returned None and pruned every branch before reaching a leaf, so deep-ε
targets returned no solution (Rz(π/64) @1e-10: None → 93 T, diamond
9.2e-11). Widen z to i128; the basis and reconstructed x stay i64.

+++ src/synthesis/lattice/omega/integer.rs
@@ -309,7 +309,7 @@ pub fn find_aligned_outcome_mpfr(
         abort,
         &node_budget,
         &truncated,
-        |z: &[i64; 8]| {
+        |z: &[i128; 8]| {
             let n_so_far = count.load(Ordering::Relaxed);
             if n_so_far >= max_leaf_checks {
                 // Leaf budget exhausted: abort the walk (the flag is the
